backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Chess analysis
python-chess==1.10.0
# stockfish==3.28.0  # Will add later

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.2.0
